vlm-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
vllm==0.2.7
torch==2.1.0
transformers==4.36.0
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
